pyproject.toml
[tool.poetry]
name = "networked-ai-system-course"
version = "0.1.0"
description = ""
authors = ["Korbinian Pfab <korbinian.pfab@siemens.com>"]
readme = "README.md"
packages = [{include = "networked_ai_system_course"}]

[tool.poetry.dependencies]
python = "3.10.12"
scikit-learn = "^1.4.0"
plotly = "^5.18.0"
tensorflow = "2.15"
paho-mqtt = "^1.6.1"
numpy = "^1.26.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
